+++ memory_linkedlist.h
@@ -26,9 +26,9 @@ struct Memory_list
 
 struct Memory_list *memlist_create(size_t memory_size);
 size_t memlist_len(struct Memory_list* lst);
-int memlist_add_circular(struct Memory_list* lst, size_t pid, size_t process_size);
-int memlist_add_worst(struct Memory_list* lst, size_t pid, size_t process_size);
-int memlist_remove_node(struct Memory_list* lst, size_t program_id);
+long long memlist_add_circular(struct Memory_list* lst, size_t pid, size_t process_size);
+long long memlist_add_worst(struct Memory_list* lst, size_t pid, size_t process_size);
+long long memlist_remove_node(struct Memory_list* lst, size_t program_id);
 int memlist_remove_node_index(struct Memory_list* lst, size_t index);
 struct Node *memlist_get_node(struct Memory_list* lst, size_t index);
 void memlist_dump(struct Memory_list* lst);
